debugging catGrade()
I found some naming conflictions. The 2 initialized strings "name" and "category" are named exactly like the vector type data members in our Gradebook object. I removed "name" because I don't think it was used in the function and I renamed "category" to "selected_category" for the sake of consistency and to avoid any confusion.

The integer variable "input" used to select the category was never assigned a value, so I just added a quick line in there to do so.

In the for loop, you had it iterate from 0 to gradebook.size(). I wasn't sure what this size was referring to and my IDE also threw an error, so I just changed it to go from 0 to id.size(). I think any other vector data member in the class would've worked too.

I was getting errors in my IDE, so I changed the way the average was calculated within the for loop. It seems like a numeric value was being divided by a string value, so I just changed it to divide by a grade counter of the coursework instead.

I changed the invalid category print statement to only print if there were no grades in the selected category.