Slightly modified readFile() and saveFile()
readFile() Modification
Previously, Gradebook's data member, "filename", is never modified to be used
elsewhere in the saveFile function. The first line in the ReadFile function now
does this.

saveFile() Modification
Previously, when std::tostring() was used on a double value, it would output a value
with 6 zeroes after the decimal place. I thought of two ways to change this.
I could've set the precision using a function from the <iomanip> library. However, 
instead, I thought taking out the tostring functions directly was easier.